Refactor clock sync. Move logic to cpp. Use an entt::delegate to generate packets. Connect packet signal in the client and server to this delegate.
