+++ CMakeLists.txt
@@ -119,6 +119,7 @@ add_library(Edyn
     src/edyn/networking/sys/update_aabbs_of_interest.cpp
     src/edyn/networking/extrapolation_job.cpp
     src/edyn/networking/packet/util/pool_snapshot.cpp
+    src/edyn/networking/util/clock_sync.cpp
     src/edyn/context/settings.cpp
     src/edyn/edyn.cpp
     src/edyn/time/common/time.cpp
